requirements.txt
# Bot framework
python-telegram-bot